+++ src/main/java/com/the2ang/cardmemory/repository/card/impl/MainCategoryRepositoryImpl.java
@@ -1,14 +1,13 @@
 package com.the2ang.cardmemory.repository.card.impl;
 
 import com.the2ang.cardmemory.entity.card.MainCategory;
-import com.the2ang.cardmemory.repository.bookstore.impl.BaseBookstoreRepositoryImpl;
 import com.the2ang.cardmemory.repository.card.MainCategoryRepository;
 import jakarta.persistence.EntityManager;
 
 import java.util.List;
 
 public class MainCategoryRepositoryImpl
-        extends BaseCardRepositoryImpl<MainCategory, Long> implements MainCategoryRepository {
+        extends BaseCardRepositoryImpl<MainCategory, Integer > implements MainCategoryRepository {
     public MainCategoryRepositoryImpl(EntityManager em) {
         super(MainCategory.class, em);
     }
